优秀的时间操作

+++ core/类型转换.go
@@ -20,8 +20,8 @@ func E到字节集(value interface{}) []byte {
 func E到字节(value interface{}) byte {
 	return gconv.Byte(value)
 }
-func E到整数(value interface{}) int {
-	return gconv.Int(value)
+func E到整数(value interface{}) int64 {
+	return gconv.Int64(value)
 }
 
 func E到数值(value interface{}) float64 {
